requirements.txt
langchain
langchain-community
langchain-openai
transformers
sentence-transformers
datasets
torch
chromadb
rank_bm25
langgraph
streamlit
gradio
python-dotenv
pydantic
scikit-learn
langchain-huggingface
langchain-chroma


streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain-huggingface
datasets
sentence-transformers
pydantic
python-dotenv
huggingface_hub

# PDF and OCR handling
pdf2image
pillow
pytesseract
unstructured[local-inference,pdf]
torch
onnxruntime

# For Amazon Textract loader (optional)
boto3
